Rework searcher - filter-sidebar, implemented text
