+++ app/components/filter-sidebar.hbs
@@ -1,0 +1,11 @@
+<aside class="au-c-sidebar au-o-layout--large" style="width: 30%;">
+  <div class="au-c-sidebar__content au-u-padding">
+    <AuContent @skin="tiny">
+      <h1>Ik zoek...</h1>
+      <section>
+        {{yield}}
+      </section>
+    </AuContent>
+  </div>
+</aside>
+
