+++ app/components/filter-sidebar/filter.hbs
@@ -3,8 +3,8 @@
       style="width: 100%;"
       class="au-u-padding-top-small au-u-padding-bottom-small"
     >
-    <AuLabel for={{@parent.id}} style="width: 30%;max-width: 30%;">
-      {{@parent.searchLabel}}
+    <AuLabel for={{@id}} style="width: 30%;max-width: 30%;">
+      {{@label}}
     </AuLabel>
     <section
       style="width: 70%;max-width: 70%; gap: 8px 0;"
