+++ app/components/filter-sidebar/select-filter.hbs
@@ -1,0 +1,23 @@
+<FilterSidebar::Filter @id={{@id}} @label={{@label}}>
+    <PowerSelect
+      style="width: 100%;"
+      id={{@id}}
+
+      @onChange={{action this.selectChange}}
+      @selected={{@selected}}
+      @options={{@options}}
+
+      @allowClear=true
+      @searchField="label"
+
+      @placeholder={{@placeholder}}
+      @noMatchesMessage={{concat "Geen " @label " gevonden"}}
+      @searchMessage="Aan het zoeken..."
+      @searchEnabled={{true}}
+
+      as |singleSelect|
+    >
+      {{singleSelect.label}}
+
+    </PowerSelect>
+</FilterSidebar::Filter>
