+++ app/components/filter-sidebar/select-filter.ts
@@ -1,0 +1,13 @@
+import Component from '@glimmer/component';
+import { action } from "@ember/object";
+import { tracked } from '@glimmer/tracking';
+
+export default class SearcherSelectFilterComponent extends Component {
+    @tracked options?: any;
+    @tracked selected?: any;
+
+    @action
+    async selectChange(m: any) {
+        this.selected = m;
+    }
+}
