+++ app/components/filter-sidebar/text-filter.hbs
@@ -1,0 +1,13 @@
+<FilterSidebar::Filter @id={{@id}} @label={{@label}}>
+    <AuInput
+        id={{@id}}
+        style="width: 100%;"
+
+        @type={{"text"}}
+        @value={{@value}}
+        placeholder={{@placeholder}}
+        onChange={{action this.onChange}}
+    />
+    {{yield}}
+
+</FilterSidebar::Filter>
