--- app/components/searcher.hbs
@@ -1,97 +1,0 @@
-<aside class="au-c-sidebar au-o-layout--large" style="width: 30%;">
-  <div class="au-c-sidebar__content au-u-padding">
-    <AuContent @skin="tiny">
-      <h1>Ik zoek...</h1>
-      <section>
-
-        {{#each this.filters as |filter|}}
-
-          {{#if filter.isSelectFilter}}
-            <Searcher::SelectFilter 
-              @info={{filter}}
-              @searcherUpdateFilter={{this.updateFilter}}
-            />
-          {{else if filter.isDateRangeFilter}}
-            <Searcher::DateRangeFilter
-              @info={{filter}}
-              @searcherUpdateFilter={{this.updateFilter}}
-            />
-          {{else}}
-            <Searcher::TextFilter
-              @info={{filter}}
-              @searcherUpdateFilter={{this.updateFilter}}
-            />
-          {{/if}}
-        {{/each}}
-      </section>
-    </AuContent>
-  </div>
-</aside>
-
-
-
-
-
-
-<main
-    class="au-u-flex au-u-flex--column"
-    style="width: 100%; height: 100%;"
-  >
-    <AuFormRow
-      @alignment="inline"
-      class="au-u-margin-bottom au-u-margin-top au-u-padding-right-large au-u-padding-left-large"
-    >
-      <section style="margin-left: auto;" class="au-u-flex">
-        <AuLabel for="sort-on" style="width: 80px; white-space: nowrap;">
-          Sorteer op
-        </AuLabel>
-        <select
-          type="select"
-          {{!--onchange={{action "handleSort"}}--}}
-          class="au-c-input"
-          style="width: 200px;"
-        >
-          <option>Relevantie</option>
-          <option>Datum</option>
-        </select>
-      </section>
-    </AuFormRow>
-    
-    <div
-      class="au-u-padding-left-large au-u-padding-right-large au-u-flex au-u-flex--column"
-      style="overflow-y: scroll; height: auto; padding-bottom:10vh; scroll-behavior: smooth;"
-    >
-
-
-      <section
-        {{did-insert this.setupInViewport}}
-        class="au-o-flow au-u-flex au-u-flex--column"
-      >
-        {{#if this.values}}
-          {{#each this.values as |item|}}
-            <DecisionCard
-              @startDate={{item.session.startedAt}}
-              @endDate={{item.session.endedAt}}
-              @item={{item}}
-            />
-          {{/each}}
-        {{else}}
-          <AuAlert @title={{"Oeps!"}} @skin={{"warning"}} @icon={{"archive"}}>
-            <p>Deze gemeente heeft spijtig genoeg geen resultaten!</p>
-          </AuAlert>
-        {{/if}}
-
-        {{#if this.values}}
-          <div
-            id="loader"
-            style="height: 80px; display: flex; justify-content: center; align-items: center;"
-          >
-            <AuLoader />
-          </div>
-        {{/if}}
-      </section>
-  </div>
-  
-</main>
-      
-{{did-insert this.request}}
